+++ History.md
@@ -7,6 +7,7 @@ Added support for Hindi (hi), although some systems may need to install fonts wh
 * Fix --dry-run option ignored when set via profile (#248 / #255 Igor Afonov)
 * More clear suggested ruby code for undefined steps (#328 / #331 @martco)
 * Fix exception in Html formatter with --expand mode and undefined steps (#336 Roberto Decurnex)
+* Fix Table.diff! problem with :surplus_row => false and interleaved surplus rows (#220)
 
 ## [v1.2.1](https://github.com/cucumber/cucumber/compare/v1.2.0...v1.2.1)